+++ datahub-web-react/src/app/ingest/source/builder/__tests__/utils.test.ts
@@ -5,7 +5,7 @@ import {
     invalidTypeRecipe,
     missingSourceRecipe,
     invalidNestedFieldRecipe,
-} from './ingestionMocks';
+} from '../ingestionMocks';
 
 describe('ingestion builder utils', () => {
     it('should return true for a valid recipe', () => {
